+++ include/list.h
@@ -6,7 +6,7 @@ struct list_head
 	struct list_head *next, *prev;
 };
 
-#define offsetof(type, member) ((unsigned int)&((type *)0)->member)
+#define offsetof(type, member) ((unsigned long)&((type *)0)->member)
 
 #define container_of(ptr, type, member) \
 		(type *)((char *)(ptr) - offsetof(type, member));
